pipeline/requirements.txt
aws_cdk.core==1.143.0
